# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Restaurant Project CRUD",
    "description": "Restaurant project API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [
    {
      "name": "Restaurants",
      "description": "API for users in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces":[
    "application/json"
  ],

  "paths": {
    "/restaurants": {
      "get" : {
        "tags": ["Restaurants"],
        "summary": "Get all restaurants in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    },
    "/restaurant": {
      "post" : {
        "tags": ["Restaurants"],
        "summary": "Create New Restaurant",
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "Restaurant we want to create",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New restaurant is created",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    },
    "/restaurant/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of restaurant to delete",
          "type": "integer"
        }
      ],
      "delete": {
        "tags":["Restaurants"],
        "summary": "Delete a restaurant",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Deleted Successfully",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Restaurant": {
      "required": ["id", "name"],
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}